+++ packages/analytics-browser/src/browser-client.ts
@@ -21,7 +21,6 @@ import {
   BrowserOptions,
   BrowserConfig,
   BrowserClient,
-  Plugin,
 } from '@amplitude/analytics-core';
 import {
   getAttributionTrackingConfig,
@@ -47,27 +46,17 @@ import { createBrowserJoinedConfigGenerator } from './config/joined-config';
 import { autocapturePlugin } from '@amplitude/plugin-autocapture-browser';
 import { WebAttribution } from './attribution/web-attribution';
 
-interface PluginHost {
-  plugin(name: string): Plugin | undefined;
-}
-
-export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient, PluginHost {
+/**
+ * Exported for `@amplitude/unified` or integration with blade plugins.
+ * If you only use `@amplitude/analytics-browser`, use `amplitude.init()` or `amplitude.createInstance()` instead.
+ */
+export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient {
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   config: BrowserConfig;
   previousSessionDeviceId: string | undefined;
   previousSessionUserId: string | undefined;
   webAttribution: WebAttribution | undefined;
-
-  plugin(name: string): Plugin | undefined {
-    const plugin = this.timeline.plugins.find((plugin) => plugin.name === name);
-    if (plugin === undefined) {
-      this.config.loggerProvider.debug(`Cannot find plugin with name ${name}`);
-      return undefined;
-    }
-
-    return plugin;
-  }
 
   init(apiKey = '', userIdOrOptions?: string | BrowserOptions, maybeOptions?: BrowserOptions) {
     let userId: string | undefined;
@@ -188,6 +177,8 @@ export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient, Pl
     this.config.loggerProvider.debug('function setUserId: ', userId);
     if (userId !== this.config.userId || userId === undefined) {
       this.config.userId = userId;
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+      this.timeline.onIdentityChanged({ userId: userId });
       setConnectorUserId(userId, this.config.instanceName);
     }
   }
@@ -202,8 +193,12 @@ export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient, Pl
       return;
     }
     this.config.loggerProvider.debug('function setDeviceId: ', deviceId);
-    this.config.deviceId = deviceId;
-    setConnectorDeviceId(deviceId, this.config.instanceName);
+    if (deviceId !== this.config.deviceId) {
+      this.config.deviceId = deviceId;
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+      this.timeline.onIdentityChanged({ deviceId: deviceId });
+      setConnectorDeviceId(deviceId, this.config.instanceName);
+    }
   }
 
   reset() {
@@ -229,6 +224,11 @@ export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient, Pl
     this.config.loggerProvider.debug('function setSessionId: ', sessionId);
 
     const previousSessionId = this.getSessionId();
+    if (previousSessionId !== sessionId) {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+      this.timeline.onSessionIdChanged(sessionId);
+    }
+
     const lastEventTime = this.config.lastEventTime;
     let lastEventId = this.config.lastEventId ?? -1;