+++ packages/plugin-page-view-tracking-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.17...@amplitude/plugin-page-view-tracking-browser@2.3.18) (2025-04-17)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
+## [2.3.17](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.16...@amplitude/plugin-page-view-tracking-browser@2.3.17) (2025-04-15)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
 ## [2.3.16](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.3.15...@amplitude/plugin-page-view-tracking-browser@2.3.16) (2025-04-10)
 
 **Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser