+++ packages/plugin-web-attribution-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.42](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-attribution-browser@2.1.41...@amplitude/plugin-web-attribution-browser@2.1.42) (2025-04-17)
+
+**Note:** Version bump only for package @amplitude/plugin-web-attribution-browser
+
+
+
+
+
+## [2.1.41](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-attribution-browser@2.1.40...@amplitude/plugin-web-attribution-browser@2.1.41) (2025-04-15)
+
+**Note:** Version bump only for package @amplitude/plugin-web-attribution-browser
+
+
+
+
+
 ## [2.1.40](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-attribution-browser@2.1.39...@amplitude/plugin-web-attribution-browser@2.1.40) (2025-04-10)
 
 **Note:** Version bump only for package @amplitude/plugin-web-attribution-browser